+++ src/components/layout/SideNav/SideNav.tsx
@@ -1,8 +1,8 @@
 'use client';
 import { MainCategory } from '@/model/category';
-import CategoryCard from './CategoryCard';
+import CategoryCard from '../../CategoryCard';
 import styles from './css/SideNav.module.css';
-import ArrowBarIcon from './icons/ArrowBarIcon';
+import ArrowBarIcon from '../../icons/ArrowBarIcon';
 import { useState } from 'react';
 
 type Props = {
